feat(core): added ability to automatically recompile when files change
Closes: #1

Added a watcher service to automatically trigger recompilation of
the source SASS directory when child file changes are detected. This
initial implementation naively recompiles all public SASS files
whenever a file change is detected. May need to make this smarter
about dependencies later if the compilation process proves to be
overly expensive.

This change also includes the creation of a changelog and a sizeable
code reorganization to move file utility functions to their own ns.

Since this change alters the way the task is invoked and introduces
a significant feature, this commit will be the 0.2.0 release.